+++ backend/src/main/java/middle_point_search/backend/domains/memberRoom/MemberRoomRepository.java
@@ -3,4 +3,6 @@ package middle_point_search.backend.domains.memberRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRoomRepository extends JpaRepository<MemberRoom, Long> {
+
+	Boolean existsByMember_IdAndRoom_Id(Long memberId, Long roomId);
 }
