+++ backend/src/main/java/middle_point_search/backend/domains/timeVoteRoom/service/TimeVoteRoomService.java
@@ -4,13 +4,7 @@ import static middle_point_search.backend.common.exception.errorCode.UserErrorCo
 import static middle_point_search.backend.domains.timeVoteRoom.dto.TimeVoteRoomDTO.*;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Comparator;
-import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 
 import org.springframework.stereotype.Service;
@@ -18,13 +12,10 @@ import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
 import middle_point_search.backend.common.exception.CustomException;
-import middle_point_search.backend.domains.member.domain.Member;
 import middle_point_search.backend.domains.room.domain.Room;
+import middle_point_search.backend.domains.room.service.RoomService;
 import middle_point_search.backend.domains.timeVoteRoom.domain.MeetingDate;
-import middle_point_search.backend.domains.timeVoteRoom.domain.TimeVote;
 import middle_point_search.backend.domains.timeVoteRoom.domain.TimeVoteRoom;
-import middle_point_search.backend.domains.timeVoteRoom.repository.MeetingDateRepository;
-import middle_point_search.backend.domains.timeVoteRoom.repository.TimeVoteRepository;
 import middle_point_search.backend.domains.timeVoteRoom.repository.TimeVoteRoomRepository;
 
 @Service
@@ -33,19 +24,21 @@ import middle_point_search.backend.domains.timeVoteRoom.repository.TimeVoteRoomR
 public class TimeVoteRoomService {
 
 	private final TimeVoteRoomRepository timeVoteRoomRepository;
-	private final TimeVoteRepository timeVoteRepository;
-	private final MeetingDateRepository meetingDateRepository;
+	private final RoomService roomService;
 
-	//시간 투표방 생성
+	// 시간 투표방 생성
 	@Transactional(rollbackFor = {CustomException.class})
-	public TimeVoteRoomCreateResponse createTimeVoteRoom(Room room, TimeVoteRoomCreateRequest request) {
+	public TimeVoteRoomCreateResponse createTimeVoteRoom(Long roomId, TimeVoteRoomCreateRequest request) {
+		boolean exists = timeVoteRoomRepository.existsByRoom_Id(roomId);
 
-		boolean exists = timeVoteRoomRepository.existsByRoom(room);
-
-		//방존재여부 확인
+		// 방존재여부 확인
 		if (exists) {
 			throw CustomException.from(DUPLICATE_VOTE_ROOM);
 		}
+
+		// 방 조회
+		Room room = roomService.findRoom(roomId)
+			.orElseThrow(() -> CustomException.from(ROOM_NOT_FOUND));
 
 		TimeVoteRoom timeVoteRoom = new TimeVoteRoom(room, request.getDates());
 		TimeVoteRoom savedTimeVoteRoom = timeVoteRoomRepository.save(timeVoteRoom);
@@ -55,15 +48,13 @@ public class TimeVoteRoomService {
 
 	//시간투표방 재생성하기
 	@Transactional(rollbackFor = {CustomException.class})
-	public TimeVoteRoomCreateResponse recreateTimeVoteRoom(Room room, TimeVoteRoomCreateRequest request) {
-
+	public TimeVoteRoomCreateResponse recreateTimeVoteRoom(Long roomId, TimeVoteRoomCreateRequest request) {
 		// 기존 투표방 삭제
-		TimeVoteRoom existingTimeVoteRoom = timeVoteRoomRepository.findByRoom(room)
-			.orElseThrow(() -> CustomException.from(VOTE_ROOM_NOT_FOUND));
+		timeVoteRoomRepository.deleteByRoom_Id(roomId);
 
-		// 먼저 투표와 관련된 모든 데이터 삭제
-		timeVoteRoomRepository.delete(existingTimeVoteRoom);
-		timeVoteRoomRepository.flush();
+		// 방 조회
+		Room room = roomService.findRoom(roomId)
+			.orElseThrow(() -> CustomException.from(ROOM_NOT_FOUND));
 
 		// 새로운 투표방 생성
 		TimeVoteRoom timeVoteRoom = new TimeVoteRoom(room, request.getDates());
@@ -72,103 +63,9 @@ public class TimeVoteRoomService {
 		return TimeVoteRoomCreateResponse.from(savedTimeVoteRoom.getId());
 	}
 
-	//시간투표하기
-	@Transactional(rollbackFor = {CustomException.class})
-	public void vote(Member member, Room room, TimeVoteRoomVoteRequest request) {
-
-		TimeVoteRoom timeVoteRoom = timeVoteRoomRepository.findByRoom(room)
-			.orElseThrow(() -> CustomException.from(VOTE_ROOM_NOT_FOUND));
-
-		boolean alreadyVoted = timeVoteRepository.existsByTimeVoteRoomAndMember(timeVoteRoom, member);
-
-		if (alreadyVoted) {
-			throw CustomException.from(ALREADY_VOTED);
-		}
-		List<TimeVote> timeVotes = createNewTimeVotes(request.getDateTime(), timeVoteRoom, member);
-		timeVoteRepository.saveAll(timeVotes);
-	}
-
-	// 시간 투표 수정
-	@Transactional(rollbackFor = {CustomException.class})
-	public void updateVote(Member member, Room room, TimeVoteRoomVoteRequest request) {
-
-		TimeVoteRoom timeVoteRoom = timeVoteRoomRepository.findByRoom(room)
-			.orElseThrow(() -> CustomException.from(VOTE_ROOM_NOT_FOUND));
-
-		List<TimeVote> existingVotes = timeVoteRepository.findAllByTimeVoteRoomAndMember(timeVoteRoom, member);
-		if (existingVotes.isEmpty()) {
-			throw CustomException.from(VOTE_NOT_FOUND);
-		}
-		// 기존 투표 삭제
-		timeVoteRepository.deleteAll(existingVotes);
-		// 새로운 투표 추가
-		List<TimeVote> timeVotes = createNewTimeVotes(request.getDateTime(), timeVoteRoom, member);
-		timeVoteRepository.saveAll(timeVotes);
-	}
-
-	private List<TimeVote> createNewTimeVotes(List<TimeRange> dateTimeRanges, TimeVoteRoom timeVoteRoom,
-		Member member) {
-		List<TimeVote> timeVotes = new ArrayList<>();
-		for (TimeRange timeRange : dateTimeRanges) {
-			LocalDateTime memberAvailableStartTime = timeRange.getMemberAvailableStartTime();
-			LocalDateTime memberAvailableEndTime = timeRange.getMemberAvailableEndTime();
-			LocalDateTime startDateTime = memberAvailableStartTime.toLocalDate().atStartOfDay();
-
-			MeetingDate meetingDate = meetingDateRepository.findByTimeVoteRoomAndDate(
-					timeVoteRoom,
-					startDateTime.toLocalDate())
-				.orElseThrow(() -> CustomException.from(CANDIDATE_NOT_FOUND));
-
-			TimeVote timeVote = new TimeVote(
-				timeVoteRoom,
-				meetingDate,
-				member,
-				memberAvailableStartTime,
-				memberAvailableEndTime);
-
-			timeVotes.add(timeVote);
-		}
-		return timeVotes;
-	}
-
-	// 시간 투표 현황 정보 조회
-	public TimeVoteRoomResultResponse getTimeVoteResult(Room room) {
-		TimeVoteRoom timeVoteRoom = timeVoteRoomRepository.findByRoom(room)
-			.orElseThrow(() -> CustomException.from(VOTE_ROOM_NOT_FOUND));
-		List<MeetingDate> meetingDates = meetingDateRepository.findByTimeVoteRoom(timeVoteRoom);
-
-		Map<String, List<TimeVoteDetail>> result = new LinkedHashMap<>();
-		meetingDates.sort(Comparator.comparing(MeetingDate::getDate));
-
-		for (MeetingDate meetingDate : meetingDates) {
-			String date = meetingDate.getDate().toString();
-			List<TimeVoteDetail> details = new ArrayList<>();
-
-			// 해당 날짜의 모든 투표 정보 가져오기
-			List<TimeVote> timeVotes = timeVoteRepository.findAllByTimeVoteRoomAndMeetingDate(timeVoteRoom,
-				meetingDate);
-			for (TimeVote vote : timeVotes) {
-				List<TimeRange> dateTimeList = Arrays.asList(
-					new TimeRange(
-						vote.getMemberAvailableStartTime(),
-						vote.getMemberAvailableEndTime()
-					)
-				);
-				TimeVoteDetail detail = TimeVoteDetail.from(vote.getMember().getName(), dateTimeList);
-				details.add(detail);
-			}
-			result.put(date, details);
-		}
-
-		List<TimeVote> distinctVotes = timeVoteRepository.findDistinctByTimeVoteRoom(timeVoteRoom);
-		int totalMemberNum = (int)distinctVotes.stream().map(TimeVote::getMember).distinct().count();
-		return TimeVoteRoomResultResponse.from(result, totalMemberNum);
-	}
-
-	//시간투표방 존재 여부 확인, 존재시 true, 존재하지 않을시 false 반환
-	public TimeVoteRoomGetResponse getTimeVoteRoom(String roomId) {
-
-		Optional<TimeVoteRoom> timeVoteRoomOptional = timeVoteRoomRepository.findByRoom_IdentityNumber(roomId);
+	// 시간투표방 조회
+	public TimeVoteRoomGetResponse findTimeVoteRoomAndMakeDTO(Long roomId) {
+		Optional<TimeVoteRoom> timeVoteRoomOptional = timeVoteRoomRepository.findByRoom_Id(roomId);
 
 		return timeVoteRoomOptional
 			.map(timeVoteRoom -> {
@@ -181,57 +78,8 @@ public class TimeVoteRoomService {
 			.orElseGet(() -> TimeVoteRoomGetResponse.from(false, null));
 	}
 
-	//먼저 시간투표방이 없다면 시간투표방없다고 에러메세지, 그 다음 시간 투표방이 있을때 투표했으면 true, 투표안했으면 false 반환
-	public VotedAndVoteItemsGetResponse getVotedAndVoteItems(Member member, Room room) {
-
-		TimeVoteRoom timeVoteRoom = timeVoteRoomRepository.findByRoom(room)
-			.orElseThrow(() -> CustomException.from(VOTE_ROOM_NOT_FOUND));
-
-		//내 시간투표 가져오기
-		List<TimeRange> myVotes = getMyVotes(timeVoteRoom, member);
-		boolean myVoteExistence = !myVotes.isEmpty();
-		myVotes = myVoteExistence ? myVotes : null;
-
-		//다른 사람 시간 투표 가져오기
-		List<TimeVotePerDate> otherVotes = getOtherVotes(timeVoteRoom, member);
-		boolean otherVotesExistence = !otherVotes.isEmpty();
-		otherVotes = otherVotesExistence ? otherVotes : null;
-
-		return VotedAndVoteItemsGetResponse.from(myVoteExistence, myVotes, otherVotesExistence, otherVotes);
-	}
-
-	private List<TimeRange> getMyVotes(TimeVoteRoom timeVoteRoom, Member member) {
-		List<TimeVote> myTimeVotes = timeVoteRepository.findAllByTimeVoteRoomAndMember(timeVoteRoom,
-			member);
-
-		return myTimeVotes.stream()
-			.map(myTimeVote -> new TimeRange(myTimeVote.getMemberAvailableStartTime(),
-				myTimeVote.getMemberAvailableEndTime()))
-			.toList();
-	}
-
-	private List<TimeVotePerDate> getOtherVotes(TimeVoteRoom timeVoteRoom, Member member) {
-		List<TimeVotePerDate> otherVotes = new ArrayList<>();
-
-		List<MeetingDate> meetingDates = timeVoteRoom.getMeetingDates();
-
-		for (MeetingDate meetingDate : meetingDates) {
-			List<TimeVotePerDate.TimeVotePerDateDetail> timeVotePerDateDetails = timeVoteRepository.findAllByTimeVoteRoomAndMeetingDateExceptMember(
-					timeVoteRoom, meetingDate, member)
-				.stream()
-				.map(otherTimeVote -> {
-					TimeRange timeRange = new TimeRange(otherTimeVote.getMemberAvailableStartTime(),
-						otherTimeVote.getMemberAvailableEndTime());
-					return TimeVotePerDate.TimeVotePerDateDetail.from(otherTimeVote.getMember().getName(), timeRange);
-				})
-				.sorted(Comparator.comparing(detail -> detail.getDateTime().getMemberAvailableStartTime()))
-				.toList();
-
-			TimeVotePerDate timeVotePerDate = TimeVotePerDate.from(meetingDate.getDate(), timeVotePerDateDetails);
-
-			otherVotes.add(timeVotePerDate);
-		}
-
-		return otherVotes;
+	// 시간 투표방, 방으로 조회
+	public Optional<TimeVoteRoom> findByRoomId(Long roomId) {
+		return timeVoteRoomRepository.findByRoom_Id(roomId);
 	}
 }
